Update `age` field in `sample` table to `SMALLINT` with value constraints (0-100), update models, repository, and tests accordingly

+++ adapter/src/database/model/sample.rs
@@ -6,7 +6,7 @@ pub(crate) struct SampleRow {
     pub id: SampleId,
     pub name: String,
     pub email: String,
-    pub age: i32,
+    pub age: i16,
 }
 
 impl TryFrom<SampleRow> for Sample {
